google-cloud-function/requirements.txt
google-cloud-storage
kubernetes
google-cloud-pubsub
google-auth